Split unit tests into separate test apps
