+++ nested_admin/tests/urls.py
@@ -1,9 +1,12 @@
 import django
+from django.conf import settings
 from django.conf.urls import include, url
 from django.contrib import admin
 
 # Explicitly import to register the admins for the test models
-import nested_admin.tests.admin
+for app in settings.INSTALLED_APPS:
+    if app.startswith('nested_admin.tests'):
+        __import__("%s.admin" % app)
 
 
 urlpatterns = [
